Add beginnings of fs module
This mimics the tokio interface with a simple in-memory file system.
There's some unnecessary cloning and cases where the file system behaves
too cleanly, which need to be have injection points (like the hold,
partition, repair methods which exist for net).

Examples here are the flush and sync_all methods. Drop currently just
calls flush and blocks the runtime until it is complete. This works
fine, but doesn't allow for any injected errors, like a file not being
flushed/synced at a critical point.

Otherwise, the AsyncRead/AsyncWrite traits currently just spit out the
entire file without yielding to the runtime, which likely needs
injection points, as well.

This commit is also missing a bunch of tests and some of the API surface
area.